debug prod

+++ torchimage/pooling/operator.py
@@ -35,7 +35,7 @@ def any_conv_1d(x: torch.Tensor, w: torch.Tensor, *, dim: int, stride: int, dila
         w = w.view(1, 1, -1)  # out_channels, in_channels/groups, input kernel width
         x = F.conv1d(x, w, bias=None, stride=stride, padding=0, dilation=dilation, groups=1)
     else:
-        x = x.view(np.prod(x.shape[:dim]), 1, x.shape[dim], -1)
+        x = x.view(np.prod(x.shape[:dim], dtype=int), 1, x.shape[dim], -1)
         w = w.view(1, 1, -1, 1)  # out_channels, in_channels/groups, input kernel height, input kernel width
         x = F.conv2d(x, w, bias=None, stride=(stride, 1), padding=0, dilation=(dilation, 1), groups=1)
